third_party/WebKit/LayoutTests/external/wpt/mixed-content/allowed/meta-csp/same-host-https/audio-tag/top-level/no-redirect/allowed.https-expected.txt

This is a testharness.js-based test.
FAIL opt_in_method: meta-csp
                                 origin: same-host-https
                                 source_scheme: https
                                 context_nesting: top-level
                                 redirection: no-redirect
                                 subresource: audio-tag
                                 expectation: allowedassert_equals: The triggered event should match 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
